Made Changes to Note that create "Setters and Getters". Also changed Note to accommodate this. Also briefly changed your sound Unit to place us back on track.
Please note that we should do the following as well:
One note, and one text.  Things to be fixed again: commenting out the interface declarations (which don't yet exist), and making midiView be declared at the interface type instead of the implementation directly, and restoring final to the fields in the MidiViewImpl.