Added getPitchInt which allows for the pitch to be converted into a integer. This is done to create the right MIDI sound associated with the note.
Further began working on the MIDI View. It's been hell however Sure it will get better.

Please note that we should do the following as well:
One note, and one text.  Things to be fixed again: commenting out the interface declarations (which don't yet exist), and making midiView be declared at the interface type instead of the implementation directly, and restoring final to the fields in the MidiViewImpl.