Added for MIDI note functionality by extending Note into MIDI note. Also began to mess with Builder for sheet music
Please note that we should do the following as well:
One note, and one text.  Things to be fixed again: commenting out the interface declarations (which don't yet exist), and making midiView be declared at the interface type instead of the implementation directly, and restoring final to the fields in the MidiViewImpl.