Builder + Tempo
Added Tempo variable to the List of Notes. This data should be held by the model and will be used often. Also created setters/getters for it.

Finish MusicBuilder that takes a text file and creates a List of Notes (Model) based on it. Please note that this needs to be tested.

Am still very confused as to why the interface for it (Built by Lerner) demands that a "Builder" is returned. For now it is returning NULL.

Please note that we should do the following as well:
One note, and one text.  Things to be fixed again: commenting out the interface declarations (which don't yet exist), and making midiView be declared at the interface type instead of the implementation directly, and restoring final to the fields in the MidiViewImpl.