Builder Tested V1.0
Began Tests on the builder. So far it can read in "Mary Had a Stupid Lamb". This has been checked slightly so far. Could check far better if the console out was finished.

Please note that we should do the following as well:
One note, and one text.  Things to be fixed again: commenting out the interface declarations (which don't yet exist), and making midiView be declared at the interface type instead of the implementation directly, and restoring final to the fields in the MidiViewImpl.