OverHaul of Note and tests
Added MIDI Functionality to note.
REMOVED MIDINote. Make sure you do this if you already pulled
Edited Note tests such that it is no longer testing for overlapped notes.

Please note that we should do the following as well:
One note, and one text.  Things to be fixed again: commenting out the interface declarations (which don't yet exist), and making midiView be declared at the interface type instead of the implementation directly, and restoring final to the fields in the MidiViewImpl.