MIDI View Working 1/2
MIDI View is working... However Instruments are not setup yet :(

Please note that we should do the following as well:
One note, and one text.  Things to be fixed again: commenting out the interface declarations (which don't yet exist), and making midiView be declared at the interface type instead of the implementation directly, and restoring final to the fields in the MidiViewImpl.