GUI  2.3 + MIDI FINISHED
Finished MIDI Fully. Is working perfect so far with no known bugs. Instruments are now working correctly

Worked on GUI and have it working for the most part

Please note that we should do the following as well:
One note, and one text.  Things to be fixed again: commenting out the interface declarations (which don't yet exist), and making midiView be declared at the interface type instead of the implementation directly, and restoring final to the fields in the MidiViewImpl.