GUI  3.0
Now have Header Made for GUI
Grid for all notes with BarLines and Bar Headers are Drawn!

Feel Free to Run the Main and take a look!

Please note that we should do the following as well:
One note, and one text.  Things to be fixed again: commenting out the interface declarations (which don't yet exist), and making midiView be declared at the interface type instead of the implementation directly, and restoring final to the fields in the MidiViewImpl.