GUI  4.0
FULL GUI

Only Minor Edit that needs to be done is to get rid of overlapping Bar Column's
Also THICKEN middle C line

Please note that we should do the following as well:
One note, and one text.  Things to be fixed again: commenting out the interface declarations (which don't yet exist), and making midiView be declared at the interface type instead of the implementation directly, and restoring final to the fields in the MidiViewImpl.